Use import/export for GGRC.infer_object_type
