+++ src/ggrc-client/js/controllers/summary_widget_controller.js
@@ -13,6 +13,9 @@ import {
   getDefaultStatesForModel,
 } from '../plugins/utils/state-utils';
 import {loadScript} from '../plugins/ggrc_utils';
+import {
+  inferObjectType,
+} from '../plugins/utils/models-utils';
 
 export default can.Control({
   defaults: {
@@ -48,7 +51,7 @@ export default can.Control({
     $(function () {
       if (GGRC.page_object) {
         $.extend(that.defaults, {
-          model: GGRC.infer_object_type(GGRC.page_object),
+          model: inferObjectType(GGRC.page_object),
           instance: GGRC.page_instance(),
         });
       }
