+++ src/ggrc-client/js/ggrc_base.js
@@ -4,6 +4,7 @@
 */
 
 import PersistentNotifier from './plugins/persistent_notifier';
+import {inferObjectType} from './plugins/utils/models-utils';
 
 (function (GGRC, moment) {
   GGRC.mustache_path = '/static/mustache';
@@ -32,61 +33,13 @@ import PersistentNotifier from './plugins/persistent_notifier';
     });
 
   $.extend(GGRC, {
-    get_object_type_decision_tree: function () {
-      let tree = {},
-        extensions = GGRC.extensions || []
-      ;
-
-      can.each(extensions, function (extension) {
-        if (extension.object_type_decision_tree) {
-          if (can.isFunction(extension.object_type_decision_tree)) {
-            $.extend(tree, extension.object_type_decision_tree());
-          } else {
-            $.extend(tree, extension.object_type_decision_tree);
-          }
-        }
-      });
-
-      return tree;
-    },
-
-    infer_object_type: function (data) {
-      let decision_tree = GGRC.get_object_type_decision_tree();
-
-      function resolve_by_key(subtree, data) {
-        let kind = data[subtree._key];
-        let model;
-        can.each(subtree, function (v, k) {
-          if (k != '_key' && v.meta_kinds.indexOf(kind) >= 0) {
-            model = v;
-          }
-        });
-        return model;
-      }
-
-      function resolve(subtree, data) {
-        if (typeof subtree === 'undefined')
-          return null;
-        return can.isPlainObject(subtree) ?
-          subtree._discriminator(data) :
-          subtree;
-      }
-
-      if (!data) {
-        return null;
-      } else {
-        return can.reduce(Object.keys(data), function (a, b) {
-          return a || resolve(decision_tree[b], data[b]);
-        }, null);
-      }
-    },
     make_model_instance: function (data) {
       if (!data) {
         return null;
       } else if (!!GGRC.page_model && GGRC.page_object === data) {
         return GGRC.page_model;
       } else {
-        return GGRC.page_model = GGRC.infer_object_type(data).model($.extend({}, data));
+        return GGRC.page_model = inferObjectType(data).model($.extend({}, data));
       }
     },
