+++ src/ggrc-client/js/plugins/utils/models-utils.js
@@ -39,6 +39,37 @@ const getModelInstance = (id, type, requiredAttr) => {
   return promise;
 };
 
+const inferObjectType = (data) => {
+  let decision_tree = _getObjectTypeDecisionTree();
+
+  function resolve_by_key(subtree, data) {
+    let kind = data[subtree._key];
+    let model;
+    can.each(subtree, function (v, k) {
+      if (k != '_key' && v.meta_kinds.indexOf(kind) >= 0) {
+        model = v;
+      }
+    });
+    return model;
+  }
+
+  function resolve(subtree, data) {
+    if (typeof subtree === 'undefined')
+      return null;
+    return can.isPlainObject(subtree) ?
+      subtree._discriminator(data) :
+      subtree;
+  }
+
+  if (!data) {
+    return null;
+  } else {
+    return can.reduce(Object.keys(data), function (a, b) {
+      return a || resolve(decision_tree[b], data[b]);
+    }, null);
+  }
+};
+
 /**
  * Check the model has Related Assessments
  * @param {String} type - model type
@@ -187,10 +218,29 @@ function _removeHandler(obj, pj) {
   return dfds;
 }
 
+function _getObjectTypeDecisionTree() {
+  let tree = {},
+    extensions = GGRC.extensions || []
+  ;
+
+  can.each(extensions, function (extension) {
+    if (extension.object_type_decision_tree) {
+      if (can.isFunction(extension.object_type_decision_tree)) {
+        $.extend(tree, extension.object_type_decision_tree());
+      } else {
+        $.extend(tree, extension.object_type_decision_tree);
+      }
+    }
+  });
+
+  return tree;
+}
+
 export {
   getModelInstance,
   hasRelatedAssessments,
   relatedAssessmentsTypes,
   resolveDeferredBindings,
   handlePendingJoins,
+  inferObjectType,
 };
